+++ XCode/DataAccessLayer/Database/SqlCe.cs
@@ -147,7 +147,7 @@ namespace XCode.DataAccessLayer
         #endregion
 
         #region 分页
-        public override SelectBuilder PageSplit(SelectBuilder builder, Int32 startRowIndex, Int32 maximumRows)
+        public override SelectBuilder PageSplit(SelectBuilder builder, Int64 startRowIndex, Int64 maximumRows)
         {
             return MSPageSplit.PageSplit(builder, startRowIndex, maximumRows, false, b => CreateSession().QueryCount(b));
         }
